feat: Refactor helper props functions out. Makes for easier testing.
